# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "aalug",
            "url": "https://github.com/aalug",
            "email": "a.a.gulczynski@gmail.com"
        }
    },
    "paths": {
...
        "/users": {
            "get": {
                "description": "Get details of the logged in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.userResponse"
                        }
                    },
                    "500": {
                        "description": "Any error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "User details",
                        "name": "CreateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.userResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "User with given email already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the logged-in user",
                "tags": [
                    "users"
                ],
                "summary": "Delete user",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    },
                    "500": {
                        "description": "Any error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update the logged-in user details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "description": "User details to update",
                        "name": "UpdateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.userResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Login user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "User credentials",
                        "name": "LoginUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.loginUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.loginUserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Incorrect password",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User with given email does not exist",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/password": {
            "patch": {
                "description": "Change / update password of the logged-in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user password",
                "parameters": [
                    {
                        "description": "Users old and new password",
                        "name": "UpdateUserPasswordRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateUserPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.updateUserPasswordResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Incorrect password",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "api.Skill": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "skill": {
                    "type": "string"
                },
                "years_of_experience": {
                    "type": "integer"
                }
            }
        },
...
        "api.createUserRequest": {
            "type": "object",
            "required": [
                "desired_industry",
                "desired_job_title",
                "desired_salary_max",
                "desired_salary_min",
                "email",
                "full_name",
                "location",
                "password"
            ],
            "properties": {
                "desired_industry": {
                    "type": "string"
                },
                "desired_job_title": {
                    "type": "string"
                },
                "desired_salary_max": {
                    "type": "integer",
                    "minimum": 0
                },
                "desired_salary_min": {
                    "type": "integer",
                    "minimum": 0
                },
                "email": {
                    "type": "string"
                },
                "experience": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Skill"
                    }
                },
                "skills_description": {
                    "type": "string"
                }
            }
        },
...
        "api.loginUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "api.loginUserResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/api.userResponse"
                }
            }
        },
...
        "api.updateUserPasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "minLength": 6
                },
                "old_password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "api.updateUserPasswordResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "api.updateUserRequest": {
            "type": "object",
            "properties": {
                "desired_industry": {
                    "type": "string"
                },
                "desired_job_title": {
                    "type": "string"
                },
                "desired_salary_max": {
                    "type": "integer"
                },
                "desired_salary_min": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "experience": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "skill_ids_to_remove": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "skills_description": {
                    "type": "string"
                },
                "skills_to_add": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Skill"
                    }
                }
            }
        },
        "api.userResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "desired_industry": {
                    "type": "string"
                },
                "desired_job_title": {
                    "type": "string"
                },
                "desired_salary_max": {
                    "type": "integer"
                },
                "desired_salary_min": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "experience": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Skill"
                    }
                },
                "skills_description": {
                    "type": "string"
                }
            }
        },
...
    }
}